Initial code refactor on Pretrained modules
