+++ texar/tf/modules/encoders/__init__.py
@@ -27,5 +27,5 @@ from texar.tf.modules.encoders.hierarchical_encoders import *
 from texar.tf.modules.encoders.transformer_encoders import *
 from texar.tf.modules.encoders.multihead_attention import *
 from texar.tf.modules.encoders.conv_encoders import *
-from texar.tf.modules.encoders.bert_encoders import *
-from texar.tf.modules.encoders.xlnet_encoders import *
+from texar.tf.modules.encoders.bert_encoder import *
+from texar.tf.modules.encoders.xlnet_encoder import *
