+++ texar/tf/modules/encoders/bert_encoder.py
@@ -20,64 +20,52 @@ from __future__ import division
 from __future__ import print_function
 
 import tensorflow as tf
-from texar.tf.core import layers
+
+from texar.tf.core.layers import get_initializer, get_layer
 from texar.tf.modules.encoders.transformer_encoders import TransformerEncoder
-from texar.tf.modules.embedders import WordEmbedder, PositionEmbedder
-from texar.tf.hyperparams import HParams
-from texar.tf.modules.pretrained.pretrained_base import PretrainedBase
-from texar.tf.modules.pretrained import bert_utils
-from texar.tf.modules.encoders import EncoderBase
+from texar.tf.modules.embedders.embedders import WordEmbedder
+from texar.tf.modules.embedders.position_embedders import PositionEmbedder
+from texar.tf.modules.encoders.encoder_base import EncoderBase
+from texar.tf.modules.pretrained.bert import PretrainedBERTMixin
 
 __all__ = [
-    "BertEncoder",
+    "BERTEncoder",
 ]
 
 
-class BertEncoder(PretrainedBase, EncoderBase):
-    """Raw BERT Transformer for encoding sequences.
+class BERTEncoder(EncoderBase, PretrainedBERTMixin):
+    r"""Raw BERT Transformer for encoding sequences.
 
     This module basically stacks
-    :class:`~texar.tf.modules.embedders.WordEmbedder`,
-    :class:`~texar.tf.modules.embedders.PositionEmbedder`,
-    :class:`~texar.tf.modules.encoders.TransformerEncoder` and a dense pooler.
-
-    This module supports the architecture first proposed
-    in `(Devlin et al.)` BERT.
+    :class:`~texar.tf.modules.WordEmbedder`,
+    :class:`~texar.tf.modules.PositionEmbedder`,
+    :class:`~texar.tf.modules.TransformerEncoder` and a dense pooler.
 
     Args:
-        pretrained_model_name (optional): a str with the name
-            of a pre-trained model to load selected in the list of:
-            `bert-base-uncased`, `bert-large-uncased`, `bert-base-cased`,
-            `bert-large-cased`, `bert-base-multilingual-uncased`,
-            `bert-base-multilingual-cased`, `bert-base-chinese`.
-            If `None`, will use the model name in :attr:`hparams`.
+        pretrained_model_name (optional): a `str`, the name
+            of pre-trained model (e.g., ``bert-base-uncased``). Please refer to
+            :class:`~texar.tf.modules.PretrainedBERTMixin` for
+            all supported models.
+            If `None`, the model name in :attr:`hparams` is used.
         cache_dir (optional): the path to a folder in which the
             pre-trained models will be cached. If `None` (default),
-            a default directory will be used.
+            a default directory (``texar_data`` folder under user's home
+            directory) will be used.
         hparams (dict or HParams, optional): Hyperparameters. Missing
             hyperparameter will be set to default values. See
-            :meth:`default_hparams` for the hyperparameter sturcture
+            :meth:`default_hparams` for the hyperparameter structure
             and default values.
-
-    .. document private functions
-    .. automethod:: _build
     """
-
-    model_name = "BERT"
 
     def __init__(self,
                  pretrained_model_name=None,
                  cache_dir=None,
                  hparams=None):
-        PretrainedBase.__init__(self, pretrained_model_name, cache_dir, hparams)
-        if self.pretrained_model_dir:
-            self._hparams = HParams(self.pretrained_model_hparams,
-                                    self._hparams.todict())
+        super(BERTEncoder, self).__init__(hparams=hparams)
+
+        self.load_pretrained_config(pretrained_model_name, cache_dir)
 
         with tf.variable_scope(self.variable_scope):
-            if self._hparams.initializer:
-                tf.get_variable_scope().set_initializer(
-                    layers.get_initializer(self._hparams.initializer))
 
             # Word embedding
             self.word_embedder = WordEmbedder(
@@ -101,20 +89,26 @@ class BertEncoder(PretrainedBase, EncoderBase):
                 kwargs_i = {"units": self._hparams.hidden_size,
                             "activation": tf.tanh}
                 layer_hparams = {"type": "Dense", "kwargs": kwargs_i}
-                self.pooler = layers.get_layer(hparams=layer_hparams)
+                self.pooler = get_layer(hparams=layer_hparams)
+
+    def reset_parameters(self):
+        with tf.variable_scope(self.variable_scope):
+            if self._hparams.initializer:
+                tf.get_variable_scope().set_initializer(
+                    get_initializer(self._hparams.initializer))
 
     @staticmethod
     def default_hparams():
         """Returns a dictionary of hyperparameters with default values.
 
-        * The encoder arch is determined by the constructor argument \
-        :attr:`pretrained_model_name` if it's specified. In this case, \
-        hparams are ignored.
-        * Otherwise, the encoder arch is determined by \
-        `hparams['pretrained_model_name']` if it's specified. All other \
-        configs in hparams are ignored.
-        * If the above two are `None`, the encoder arch is defined by \
-        the configs in hparams and weights are randomly initialized.
+        * The encoder arch is determined by the constructor argument
+          :attr:`pretrained_model_name` if it's specified. In this case,
+          `hparams` are ignored.
+        * Otherwise, the encoder arch is determined by
+          `hparams['pretrained_model_name']` if it's specified. All other
+          configurations in `hparams` are ignored.
+        * If the above two are `None`, the encoder arch is defined by the
+          configurations in `hparams` and weights are randomly initialized.
 
         .. code-block:: python
 
@@ -176,49 +170,46 @@ class BertEncoder(PretrainedBase, EncoderBase):
                 'name': 'bert_encoder'
             }
 
-
-
         Here:
 
         The default parameters are values for uncased BERT-Base model.
 
+        `"pretrained_model_name"`: str or None
+            The name of the pre-trained BERT model. If None, the model
+            will be randomly initialized.
 
-        "pretrained_model_name": str or None
-             The name of the pretrained bert model. If None, the model
-             will be randomly initialized.
-
-        "embed": dict
+        `"embed"`: dict
             Hyperparameters for word embedding layer.
 
-        "vocab_size": int
-            The vocabulary size of `inputs` in `BertModel`.
+        `"vocab_size"`: int
+            The vocabulary size of `inputs` in BERT model.
 
-        "segment_embed": dict
+        `"segment_embed"`: dict
             Hyperparameters for segment embedding layer.
 
-        "type_vocab_size": int
+        `"type_vocab_size"`: int
             The vocabulary size of the `segment_ids` passed into `BertModel`.
 
-        "position_embed": dict
+        `"position_embed"`: dict
             Hyperparameters for position embedding layer.
 
-        "position_size":  int
+        `"position_size"`: int
             The maximum sequence length that this model might ever be used with.
 
-        "encoder": dict
+        `"encoder"`: dict
             Hyperparameters for the TransformerEncoder.
             See :func:`~texar.tf.modules.TransformerEncoder.default_harams`
             for details.
 
-        "hidden_size": int
+        `"hidden_size"`: int
             Size of the pooler dense layer.
 
-        "initializer": dict, optional
+        `"initializer"`: dict, optional
             Hyperparameters of the default initializer that initializes
             variables created in this module.
             See :func:`~texar.tf.core.get_initializer` for details.
 
-        "name": str
+        `"name"`: str
             Name of the module.
         """
 
@@ -318,7 +309,7 @@ class BertEncoder(PretrainedBase, EncoderBase):
 
                 - :attr:`pooled_output`: A Tensor of size \
                 `[batch_size, hidden_size]` which is the output of a \
-                pooler pretrained on top of the hidden state associated \
+                pooler berts on top of the hidden state associated \
                 to the first character of the input (`CLS`), see BERT's \
                 paper.
         """
@@ -351,8 +342,6 @@ class BertEncoder(PretrainedBase, EncoderBase):
             self._add_internal_trainable_variables()
             self._built = True
 
-            if self.pretrained_model_dir:
-                bert_utils.init_bert_checkpoint(self.pretrained_model_dir,
-                                                self.variable_scope.name)
+            self.init_pretrained_weights(self.variable_scope.name)
 
         return output, pooled_output
