+++ texar/tf/modules/pretrained/__init__.py
@@ -15,6 +15,6 @@
 Pre-trained modules of Texar library.
 """
 
-from texar.tf.modules.pretrained.bert_utils import *
 from texar.tf.modules.pretrained.pretrained_base import *
-from texar.tf.modules.pretrained.xlnet_utils import *
+from texar.tf.modules.pretrained.bert import *
+from texar.tf.modules.pretrained.xlnet import *
